[ios][fix-it] Fix NonModalEGTest tests
The tests was failing flakily because +clearDefaultBrowserPromoData did
not clear all the information stored in the NSUserDefaults. Instead of
duplicating the list in ChromeEarlGreyAppInterface and in the unit test
of default_browser_utils.h, export an test helper to clear all of them.

Refactor default_browser_utils.h to store all its data in a dictionary
stored in NSUserDefaults (ideally this should be stored in the LocalState
found from ApplicationContext but fixing this is outside of this fix).
It allows having to clear only one key from NSUserDefaults in the tests
which should be more robust.

Refactor the implementation of default_browser_utils.h to use base::Time
API instead of NSDate API for more clarity (the API to compare time with
base::Time is much more natural than the NSDate API).

Fixed: 1218866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083427
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080295}

+++ ios/chrome/browser/ui/default_promo/default_browser_promo_non_modal_metrics_util.h
@@ -32,7 +32,7 @@ enum class NonModalPromoTriggerType {
 // promo.
 void LogNonModalPromoAction(NonModalPromoAction action,
                             NonModalPromoTriggerType type,
-                            int impression_number);
+                            NSInteger impression_number);
 
 // Logs the time a non modal promo was on screen.
 void LogNonModalTimeOnScreen(base::TimeTicks initial_time);
